Blog-main/requirements.txt
gdshortener
selenium
google-generativeai
# pywin32
pillow
pyautogui
torch
torchvision
transformers
timm
kornia
einops